Extract most generators into a "generators" file
