+++ src/index/CMakeLists.txt
@@ -1,7 +1,7 @@
 sourcemeta_executable(NAMESPACE sourcemeta PROJECT registry NAME index 
   FOLDER "Registry/Index" 
-  SOURCES index.cc explorer.h toc.h configure.h.in html_partials.h html_safe.h
-    configuration.h output.h validator.h
+  SOURCES index.cc output.h generators.h  semver.h
+    configure.h.in html_partials.h html_safe.h validator.h
     "${PROJECT_SOURCE_DIR}/schemas/configuration.json")
 
 set_target_properties(sourcemeta_registry_index PROPERTIES OUTPUT_NAME sourcemeta-registry-index)
