+++ test/e2e/nav.hurl
@@ -8,6 +8,8 @@ Content-Type: application/json
 [Asserts]
 header "ETag" exists
 header "Last-Modified" exists
+jsonpath "$.url" == "/"
+jsonpath "$.canonical" == "{{base}}"
 jsonpath "$.breadcrumb" count == 0
 jsonpath "$.entries" count == 4
 jsonpath "$.entries[0].name" == "same"
@@ -38,6 +40,8 @@ Content-Type: application/json
 [Asserts]
 header "ETag" exists
 header "Last-Modified" exists
+jsonpath "$.url" == "/example"
+jsonpath "$.canonical" == "{{base}}/example"
 jsonpath "$.breadcrumb" count == 1
 jsonpath "$.breadcrumb[0].name" == "example"
 jsonpath "$.breadcrumb[0].url" == "/example"
@@ -70,6 +74,8 @@ Content-Type: application/json
 [Asserts]
 header "ETag" exists
 header "Last-Modified" exists
+jsonpath "$.url" == "/example/v2.0"
+jsonpath "$.canonical" == "{{base}}/example/v2.0"
 jsonpath "$.breadcrumb" count == 2
 jsonpath "$.breadcrumb[0].name" == "example"
 jsonpath "$.breadcrumb[0].url" == "/example"
@@ -89,6 +95,8 @@ Content-Type: application/json
 [Asserts]
 header "ETag" exists
 header "Last-Modified" exists
+jsonpath "$.url" == "/example/v2.0"
+jsonpath "$.canonical" == "{{base}}/example/v2.0"
 jsonpath "$.breadcrumb" count == 2
 jsonpath "$.breadcrumb[0].name" == "example"
 jsonpath "$.breadcrumb[0].url" == "/example"
@@ -109,6 +117,8 @@ Content-Type: application/json
 [Asserts]
 header "ETag" exists
 header "Last-Modified" exists
+jsonpath "$.url" == "/example/v2.0"
+jsonpath "$.canonical" == "{{base}}/example/v2.0"
 jsonpath "$.breadcrumb" count == 2
 jsonpath "$.breadcrumb[0].name" == "example"
 jsonpath "$.breadcrumb[0].url" == "/example"
