+++ test/schemas/configuration/fail.test.json
@@ -571,21 +571,6 @@
       }
     },
     {
-      "description": "empty analytics",
-      "valid": false,
-      "data": {
-        "url": "https://www.example.com",
-        "port": 8000,
-        "analytics": {},
-        "schemas": {
-          "example": {
-            "base": "https://schemas.example.com",
-            "path": "./schemas"
-          }
-        }
-      }
-    },
-    {
       "description": "non-string default dialect",
       "valid": false,
       "data": {
